requirements.txt
# Core Dependencies
# Requires Python 3.11 or higher

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
requests>=2.31.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Data Sources
nfl-data-py>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework & API
flask>=3.0.0
flask-cors>=4.0.0
streamlit>=1.28.0

# Storage
# sqlite3 (included in Python standard library)

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

